+++ inst/ShinyItemAnalysis/ui/uiReferences.R
@@ -286,6 +286,13 @@ uiReferences <-
                 CRC Press.
                 </li>
 
+                <li>Bartos, F., & Martinkova, P. (2024).
+                 Assessing quality of selection procedures: Lower bound of false positive rate as a function of inter-rater reliability.
+                <i>British Journal of Mathematical and Statistical Psychology, 77</i>(3), 651–671.
+                <a href = "https://doi.org/10.1111/bmsp.12343",
+                target = "_blank">See online.</a>
+                </li>
+
                 <li>Bock, R. D. (1972).
                 Estimating item parameters and latent ability when responses are
                 scored in two or more nominal categories.
@@ -340,6 +347,13 @@ uiReferences <-
                 Statistical inference for coefficient alpha.
                 <i>Applied Psychological Measurement 11</i>(1), 93-103.
                 <a href = "http://journals.sagepub.com/doi/abs/10.1177/014662168701100107",
+                target = "_blank">See online.</a>
+                </li>
+
+                <li>Hladka, A., Martinkova, P., & Brabec, M. (2024).
+                New iterative algorithms for estimation of item functioning.
+                <i>Journal of Educational and Behavioral Statistics,</i> Accepted.
+                <a href = "https://doi.org/10.48550/arXiv.2302.12648",
                 target = "_blank">See online.</a>
                 </li>
 
@@ -430,6 +444,13 @@ uiReferences <-
                 <i>Computational Aspects of Psychometric Methods: With R.</i> (1st ed.). Chapman and Hall/CRC.
                 <a href = "https://www.taylorfrancis.com/books/mono/10.1201/9781003054313",
                 target = "_blank">See online</a>
+                </li>
+
+                <li>Martinkova, P., Bartos, F., & Brabec, M. (2023).
+                Assessing inter-rater reliability with heterogeneous variance components models: Flexible approach accounting for contextual variables.
+                <i>Journal of Educational and Behavioral Statistics, 48</i>(3), 349–383.
+                <a href = "https://doi.org/10.3102/10769986221150517",
+                target = "_blank">See online.</a>
                 </li>
 
                 <li>Masters, G. N. (1982).